requirements.txt
pytorch==1.6.0
numpy==1.18.2
scipy==1.4.1
librosa==0.7.2
inflect==0.2.5
numba==0.48.0